refactor(frontend): split oversized analysis pages per project convention (APOSD)
ReturnsAnalysisPage (537 lines) and SpendingAnalysisPage (456) crammed
data fetching, derived-state calc, and JSX into one file, violating the
repo's >300-line split convention and APOSD's single-responsibility.

- ReturnsAnalysis: pure helpers -> returnsAnalysisUtils.ts, data + memo
  orchestration -> useReturnsAnalysis.ts; page is now ~390 lines of JSX.
- SpendingAnalysis: 50/30/20 calc -> spendingAnalysisUtils.ts, BudgetRuleCard
  -> components/, orchestration -> useSpendingAnalysis.ts; page now ~260.
- Fixed min-h-screen -> min-h-dvh in both (repo rule).

No behavior change -- type-check, lint, 177 tests, build all pass.
