+++ frontend/src/pages/spending-analysis/SpendingAnalysisPage.tsx
@@ -1,17 +1,11 @@
 import { motion } from 'framer-motion'
-import { useSearchParams } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 import { SCROLL_FADE_UP } from '@/constants/animations'
 import { TrendingDown, Tag, PieChart, ShieldCheck, Sparkles, PiggyBank, Activity } from 'lucide-react' // Activity used for Monthly Avg card
 import MetricCard from '@/components/shared/MetricCard'
-import { useTransactions } from '@/hooks/api/useTransactions'
-import { usePreferences } from '@/hooks/api/usePreferences'
-import { useMemo } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { formatCurrency, formatPercent } from '@/lib/formatters'
+import { formatCurrency } from '@/lib/formatters'
 import { PieChart as RechartsPie, Pie, Cell, Tooltip } from 'recharts'
-import { calculateSpendingBreakdown, SPENDING_TYPE_COLORS } from '@/lib/preferencesUtils'
-import { useAnalyticsTimeFilter } from '@/hooks/useAnalyticsTimeFilter'
-import { filterTransactionsByDateRange, computeCategoryBreakdown } from '@/lib/transactionUtils'
+import { SPENDING_TYPE_COLORS } from '@/lib/preferencesUtils'
 import EmptyState from '@/components/shared/EmptyState'
 import { FilterBanner } from '@/components/shared/FilterBanner'
 import { PageSkeleton } from '@/components/shared/LoadingSkeleton'
@@ -25,218 +19,28 @@ import {
   CohortSpendingAnalysis,
 } from '@/components/analytics'
 import { chartTooltipProps, PageHeader, ChartContainer, shouldAnimate } from '@/components/ui'
-import { SEMANTIC_COLORS } from '@/constants/chartColors'
 
-// Color for Savings (semantic, distinct from income green)
-const SAVINGS_COLOR = SEMANTIC_COLORS.savings
-
-/** Build chart data for the 50/30/20 spending breakdown */
-function buildSpendingChartData(
-  spendingBreakdown: { essential: number; discretionary: number } | null,
-  totalIncome: number,
-  savings: number,
-) {
-  if (!spendingBreakdown || totalIncome <= 0) return []
-  return [
-    { name: 'Needs', value: spendingBreakdown.essential, color: SPENDING_TYPE_COLORS.essential },
-    { name: 'Wants', value: spendingBreakdown.discretionary, color: SPENDING_TYPE_COLORS.discretionary },
-    { name: 'Savings', value: savings, color: SAVINGS_COLOR },
-  ].filter((d) => d.value > 0)
-}
-
-/** Calculate the budget rule metrics (50/30/20) based on income breakdown */
-function computeBudgetRuleMetrics(
-  spendingBreakdown: { essential: number; discretionary: number } | null,
-  totalIncome: number,
-  savings: number,
-  needsTarget: number,
-  wantsTarget: number,
-  savingsTargetPct: number,
-): {
-  essentialPercent: number
-  discretionaryPercent: number
-  savingsPercent: number
-  essentialTarget: number
-  discretionaryTarget: number
-  savingsTarget: number
-  isOverspendingEssential: boolean
-  isOverspendingDiscretionary: boolean
-  isUnderSaving: boolean
-} | null {
-  if (!spendingBreakdown || totalIncome <= 0) return null
-
-  const essentialPercent = (spendingBreakdown.essential / totalIncome) * 100
-  const discretionaryPercent = (spendingBreakdown.discretionary / totalIncome) * 100
-  const savingsPercent = (savings / totalIncome) * 100
-
-  return {
-    essentialPercent,
-    discretionaryPercent,
-    savingsPercent,
-    essentialTarget: needsTarget,
-    discretionaryTarget: wantsTarget,
-    savingsTarget: savingsTargetPct,
-    isOverspendingEssential: essentialPercent > needsTarget + 5,
-    isOverspendingDiscretionary: discretionaryPercent > wantsTarget + 5,
-    isUnderSaving: savingsPercent < savingsTargetPct - 5,
-  }
-}
-
-/** A single budget rule card (Needs/Wants/Savings) */
-function BudgetRuleCard({ title, subtitle, icon: Icon, value, percent, target, isOverBudget, accentColor, bgClass, iconBgClass, textClass, delay }: Readonly<{
-  title: string
-  subtitle: string
-  icon: React.ComponentType<{ className?: string }>
-  value: number
-  percent: number
-  target: string
-  isOverBudget: boolean
-  accentColor: string
-  bgClass: string
-  iconBgClass: string
-  textClass: string
-  delay: number
-}>) {
-  const barColor = isOverBudget ? SEMANTIC_COLORS.expense : accentColor
-  const statusColorClass = isOverBudget ? 'text-app-red' : 'text-app-green'
-
-  return (
-    <div className={`p-4 rounded-lg ${bgClass}`}>
-      <div className="flex items-center gap-3 mb-3">
-        <div className={`p-2 ${iconBgClass} rounded-lg`}>
-          <Icon className={`w-5 h-5 ${textClass}`} />
-        </div>
-        <div>
-          <p className="font-medium text-white">{title}</p>
-          <p className="text-xs text-muted-foreground">{subtitle}</p>
-        </div>
-      </div>
-      <p className={`text-2xl font-bold ${textClass} mb-2`}>
-        {formatCurrency(value)}
-      </p>
-      <div className="space-y-2">
-        <div className="flex justify-between text-sm">
-          <span className="text-muted-foreground">Current</span>
-          <span className={statusColorClass}>
-            {formatPercent(percent)}
-          </span>
-        </div>
-        <div className="h-2 bg-surface-dropdown rounded-full overflow-hidden">
-          <motion.div
-            className="h-full rounded-full"
-            initial={{ width: 0 }}
-            animate={{ width: `${Math.min(percent, 100)}%` }}
-            transition={{ duration: 0.8, ease: 'easeOut', delay }}
-            style={{ backgroundColor: barColor }}
-          />
-        </div>
-        <p className="text-xs text-text-tertiary">Target: {target} of income</p>
-      </div>
-    </div>
-  )
-}
+import { SAVINGS_COLOR } from './spendingAnalysisUtils'
+import { BudgetRuleCard } from './components/BudgetRuleCard'
+import { useSpendingAnalysis } from './useSpendingAnalysis'
 
 export default function SpendingAnalysisPage() {
   const navigate = useNavigate()
-  const [searchParams, setSearchParams] = useSearchParams()
-  const categoryFilter = searchParams.get('category')
-  const clearCategoryFilter = () => {
-    const next = new URLSearchParams(searchParams)
-    next.delete('category')
-    setSearchParams(next, { replace: true })
-  }
-  const { data: transactions } = useTransactions()
-  const { data: preferences } = usePreferences()
-  const { dateRange, timeFilterProps } = useAnalyticsTimeFilter(transactions)
-  // Convert null to undefined for component compatibility
-  const dateRangeCompat = { start_date: dateRange.start_date ?? undefined, end_date: dateRange.end_date ?? undefined }
-
-  // Filter transactions by date range, then by the category query param
-  // (set when the user lands here via a deep link from a donut slice).
-  const filteredTransactions = useMemo(() => {
-    const byDate = filterTransactionsByDateRange(transactions, dateRange)
-    if (!categoryFilter) return byDate
-    return byDate.filter((t) => t.category === categoryFilter)
-  }, [transactions, dateRange, categoryFilter])
-
-  // Calculate totals for filtered period
-  const totalSpending = useMemo(() => {
-    return filteredTransactions
-      .filter((t) => t.type === 'Expense')
-      .reduce((sum, t) => sum + Math.abs(t.amount), 0)
-  }, [filteredTransactions])
-
-  // Calculate total income for filtered period
-  const totalIncome = useMemo(() => {
-    return filteredTransactions
-      .filter((t) => t.type === 'Income')
-      .reduce((sum, t) => sum + Math.abs(t.amount), 0)
-  }, [filteredTransactions])
-
-  // Calculate savings (Income - Expenses)
-  const savings = Math.max(0, totalIncome - totalSpending)
-
-  // Get category breakdown for filtered transactions
-  const categoryBreakdown = useMemo(
-    () => computeCategoryBreakdown(filteredTransactions),
-    [filteredTransactions]
-  )
-
-  const categoriesCount = Object.keys(categoryBreakdown).length
-  const subcategoriesCount = useMemo(() => {
-    if (!filteredTransactions) return 0
-    const subs = new Set<string>()
-    filteredTransactions.filter((t) => t.type === 'Expense' && t.subcategory).forEach((t) => subs.add(`${t.category}::${t.subcategory}`))
-    return subs.size
-  }, [filteredTransactions])
-  const topCategoryEntry = Object.entries(categoryBreakdown).sort((a, b) => b[1] - a[1])[0]
-  const topCategory = topCategoryEntry?.[0] || 'N/A'
-  const topCategoryAmount = topCategoryEntry?.[1] ?? 0
-
-  // Calculate essential vs discretionary spending
-  const spendingBreakdown = useMemo(() => {
-    if (!filteredTransactions || !preferences) return null
-    return calculateSpendingBreakdown(filteredTransactions, preferences.essential_categories)
-  }, [filteredTransactions, preferences])
-
-  // Monthly average spending
-  const monthlyAvgSpending = useMemo(() => {
-    if (!filteredTransactions) return 0
-    const expenses = filteredTransactions.filter((t) => t.type === 'Expense')
-    if (expenses.length === 0) return 0
-    const months = new Set(expenses.map((t) => t.date.slice(0, 7)))
-    const total = expenses.reduce((s, t) => s + Math.abs(t.amount), 0)
-    return months.size > 0 ? total / months.size : 0
-  }, [filteredTransactions])
-
-  // Prepare spending breakdown chart data (50/30/20 rule with income base)
-  const spendingChartData = useMemo(
-    () => buildSpendingChartData(spendingBreakdown, totalIncome, savings),
-    [spendingBreakdown, savings, totalIncome]
-  )
-
-  // Spending rule targets from preferences (configurable Needs/Wants/Savings)
-  const needsTarget = preferences?.needs_target_percent ?? 50
-  const wantsTarget = preferences?.wants_target_percent ?? 30
-  const savingsTarget = preferences?.savings_target_percent ?? 20
-
-  // Calculate spending rule metrics (based on income)
-  const budgetRuleMetrics = useMemo(() => {
-    return computeBudgetRuleMetrics(spendingBreakdown, totalIncome, savings, needsTarget, wantsTarget, savingsTarget)
-  }, [spendingBreakdown, totalIncome, savings, needsTarget, wantsTarget, savingsTarget])
-
-  // Precomputed style objects for chart legend (stable references across renders)
-  const spendingLegendColorStyles = useMemo(
-    () => spendingChartData.map((item) => ({ backgroundColor: item.color })),
-    [spendingChartData]
-  )
-
-  const isLoading = !transactions
+  const {
+    categoryFilter, clearCategoryFilter,
+    timeFilterProps, dateRangeCompat, isLoading,
+    totalSpending, monthlyAvgSpending, savings,
+    categoryBreakdown, categoriesCount, subcategoriesCount,
+    topCategory, topCategoryAmount,
+    spendingBreakdown, spendingChartData, spendingLegendColorStyles,
+    budgetRuleMetrics,
+    needsTarget, wantsTarget, savingsTarget,
+  } = useSpendingAnalysis()
 
   if (isLoading) return <PageSkeleton />
 
   return (
-    <div className="min-h-screen p-4 md:p-6 lg:p-8">
+    <div className="min-h-dvh p-4 md:p-6 lg:p-8">
       <div className="max-w-7xl mx-auto space-y-6 md:space-y-8">
         <PageHeader
           title="Spending Analysis"
